Adds WeatherList component and connects API
To include actual wheather data openweather has been connected.
For now the API Key is part of the config, but we will make it
a environment variable next and update the key, so that it cannot
be misused.
In addition redux-promise is used to automatically handle promises
as action payloads.